first draft for migration on bolt Slack API
Bolt being the official Slack API, let's use it. It will make
evolution of this bot easier regarding API changes. This change
introduces the usage of Bolt and remove the usage of simpleslackapi.

A lot of TODO remains, mostly to handle errors and when a channel has
to be search by name.